% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_artist_songs_radar_chart.r
\name{create_artist_songs_radar_chart}
\alias{create_artist_songs_radar_chart}
\title{Create a radar chart of track features}
\usage{
create_artist_songs_radar_chart(
  queries = NULL,
  songs = NULL,
  vars = c(),
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{queries}{\itemize{
\item A vector of track names.  Deafualts to NULL.  Only use this or ids
}}

\item{songs}{\itemize{
\item A vector of Spotify track ids.  Deafualts to NULL.  Only use this or queries
}}

\item{vars}{\itemize{
\item A vector of variables returned from get_track_audio_features()
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A radar chart displaying valence, energy, and speechiness, along with any other inputed variables
}
\description{
Create a radar chart of track features
}
\examples{
\dontrun{
 create_artist_songs_radar_chart(songs = c("6YbhspuOar1D9WSSnfe7ds", "5Tbpp3OLLClPJF8t1DmrFD", "2NBQmPrOEEjA8VbeWOQGxO"), vars = "liveness")
 create_artist_songs_radar_chart(queries = c("Natalie", "When I Was Your Man"), vars = c("liveness", "danceability"))
}
}
