% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_artist_track_popularity_histogram.r
\name{create_artist_track_popularity_histogram}
\alias{create_artist_track_popularity_histogram}
\title{Create a histogram of artist's track popularity}
\usage{
create_artist_track_popularity_histogram(
  query = NULL,
  id = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{query}{\itemize{
\item A single artist name.  Deafualts to NULL.  Only use this or id
}}

\item{id}{\itemize{
\item A single Spotify artist id.  Deafualts to NULL.  Only use this or query
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A histogram displaying each tracks' popularity from an artist
}
\description{
Create a histogram of artist's track popularity
}
\examples{
\dontrun{
 create_artist_track_popularity_histogram(id = "2YZyLoL8N0Wb9xBt1NhZWg")
 create_artist_track_popularity_histogram(query = "Bruno Mars")
}
}
