% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_artist_track_valence_energy_visual.r
\name{create_artist_track_valence_energy_visual}
\alias{create_artist_track_valence_energy_visual}
\title{Create a valenece and energy plot based on an artist's tracks}
\usage{
create_artist_track_valence_energy_visual(
  queries = NULL,
  artist_ids = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{queries}{\itemize{
\item A String vector of artist names, taken from the search_spotify() function.  Defaults to NULL.  Only use this or ids
}}

\item{artist_ids}{\itemize{
\item A vector of Spotify artist ids.  Defaults to NULL.  Only use this or queries
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A valence and energy plot which plots the top tracks from multiple artists
}
\description{
Create a valenece and energy plot based on an artist's tracks
}
\examples{
\dontrun{
 create_artist_track_valence_energy_visual(artist_ids = "0du5cEVh5yTK9QJze8zA0C")
 create_artist_track_valence_energy_visual(artist_ids = c("0du5cEVh5yTK9QJze8zA0C", "73sIBHcqh3Z3NyqHKZ7FOL"))
 create_artist_track_valence_energy_visual(queries = c("Bruno Mars", "Anderson .Paak"))
}
}
