% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_track_bar_confidence_plot.r
\name{create_track_bar_confidence_plot}
\alias{create_track_bar_confidence_plot}
\title{Create a line chart of track bar confidence}
\usage{
create_track_bar_confidence_plot(
  query = NULL,
  id = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{query}{\itemize{
\item A single of track name.  Deafualts to NULL.  Only use this or id
}}

\item{id}{\itemize{
\item A single Spotify track id.  Deafualts to NULL.  Only use this or query
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A line chart displaying track bar confidence over the duration of the song
}
\description{
Create a line chart of track bar confidence
}
\examples{
\dontrun{
 create_track_bar_confidence_plot(id = "6YbhspuOar1D9WSSnfe7ds")
 create_track_bar_confidence_plot(query = "Natalie")
}
}
