% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_violin_plot_album_variable.r
\name{create_violin_plot_album_variable}
\alias{create_violin_plot_album_variable}
\title{Create a violin plot based on an artist's albums}
\usage{
create_violin_plot_album_variable(
  query = NULL,
  artist_id = NULL,
  limit = 7,
  var = "danceability",
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{query}{\itemize{
\item A single artist name.  Defaults to NULL.  Only put this or id
}}

\item{artist_id}{\itemize{
\item A single Spotify artist id.  Defaults to NULL.  Only put this or query
}}

\item{limit}{\itemize{
\item Number of albums wanted to return.  Valid if between 1 and 50.  Defaults to 7
}}

\item{var}{\itemize{
\item A single variable returned from the get_artist_audio_features() function.  Defaults to "danceability"
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A violin plot displaying a specific number of an artist's albums' danceability
}
\description{
Create a violin plot based on an artist's albums
}
\examples{
\dontrun{
     create_violin_plot_album_variable(query = "Bruno Mars", var = "energy")
     create_violin_plot_album_variable(artist_id = "0du5cEVh5yTK9QJze8zA0C", limit = 5)
}
}
