% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_violin_plot_artist_variable.r
\name{create_violin_plot_artist_variable}
\alias{create_violin_plot_artist_variable}
\title{Create a violin plot based on an artist's tracks}
\usage{
create_violin_plot_artist_variable(
  queries = NULL,
  artist_ids = NULL,
  var = "danceability",
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{queries}{\itemize{
\item A String vector of artist names, taken from the search_spotify() function.  Defaults to NULL.  Only use this or ids
}}

\item{artist_ids}{\itemize{
\item A vector of Spotify artist ids.  Defaults to NULL.  Only use this or queries
}}

\item{var}{\itemize{
\item A single variable returned from the get_artist_audio_features() function.  Defaults to "danceability"
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A violin plot displaying a number of artist's tracks' danceability
}
\description{
Create a violin plot based on an artist's tracks
}
\examples{
\dontrun{
 create_violin_plot_artist_variable(artist_ids = "0du5cEVh5yTK9QJze8zA0C", var = "valence")
 create_violin_plot_artist_variable(queries = c("Bruno Mars", "Anderson .Paak"))
}
}
