% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_violin_plot_genre_variable.r
\name{create_violin_plot_genre_variable}
\alias{create_violin_plot_genre_variable}
\title{Create a violin plot based on genres}
\usage{
create_violin_plot_genre_variable(
  genres,
  var = "danceability",
  limit = 20,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{genres}{\itemize{
\item A vector of Spotify genres
}}

\item{var}{\itemize{
\item A single variable returned from the get_artist_audio_features() function.  Defaults to "danceability"
}}

\item{limit}{\itemize{
\item Optional.  Number of songs wanted to return.  Valid if between 1 and 50.  Defaults to 20
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A violin plot displaying a specific number of an genres's tracks' danceability
}
\description{
Create a violin plot based on genres
}
\examples{
\dontrun{
 create_violin_plot_album_danceability(c("rap", "jazz"))
 create_violin_plot_album_danceability("hip-hop", limit = 30)
}
}
