% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album_functions.r
\name{get_album_summary}
\alias{get_album_summary}
\title{Search for album's feature summary}
\usage{
get_album_summary(
  query = NULL,
  id = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{query}{\itemize{
\item A single of album name, taken from the search_spotify() function.  Defaults to NULL.  Only use this or id
}}

\item{id}{\itemize{
\item A single Spotify album id.  Defaults to NULL.  Only use this or query
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of album data, including album name and id, album popularity, number of tracks, and the mean and standard deviation for:
danceability
energy
loudness
speechiness
acousticness
instrumentalness
liveness
valence
tempo
duration_ms
mode
}
\description{
Search for album's feature summary
}
\examples{
\dontrun{
get_album_summary(id = "1uyf3l2d4XYwiEqAb7t7fX")
get_album_summary(query = "An Evening With Silk Sonic")
}
}
