% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album_functions.r, R/tracks_functions.r
\name{get_album_track_features}
\alias{get_album_track_features}
\title{Search for Spotify album track features}
\usage{
get_album_track_features(
  query = NULL,
  id = NULL,
  authorization = get_spotify_access_token()
)

get_album_track_features(
  query = NULL,
  id = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{query}{\itemize{
\item A single album name.  Deafualts to NULL.  Only use this or id
}}

\item{id}{\itemize{
\item A single Spotify album id.  Deafualts to NULL.  Only use this or query
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of album data, including album name and id, album popularity, number of tracks, and the mean and standard deviation for:
danceability
energy
loudness
mode
speechiness
acousticness
instrumentalness
liveness
valence
tempo
duration_ms
time_signature

A dataframe of album track data, including ids and names for the tracks, whether the track is explicit, track number in the album, artist names and ids, and the following variables:
danceability
energy
key
loudness
mode
speechiness
acousticness
instrumentalness
liveness
valence
tempo
duration_ms
time_signature
}
\description{
Search for Spotify album track features

Search for Spotify album's track features
}
\examples{
\dontrun{
get_album_track_features(id = "1uyf3l2d4XYwiEqAb7t7fX")
get_album_track_features(query = "An Evening With Silk Sonic")
}
\dontrun{
get_album_track_features(id = "58ufpQsJ1DS5kq4hhzQDiI")
get_album_track_features(query = "An Evening With Silk Sonic")
}
}
