% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.r
\name{get_album_track_loudness_instrumentalness_tempo}
\alias{get_album_track_loudness_instrumentalness_tempo}
\title{Search for Spotify album track loudness and tempo}
\usage{
get_album_track_loudness_instrumentalness_tempo(
  query = NULL,
  id = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{query}{\itemize{
\item A single of album name, taken from the search_spotify() function.  Defaults to NULL.  Only use this or id
}}

\item{id}{\itemize{
\item A single Spotify album id.  Defaults to NULL.  Only use this or query
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of album track data, including album name, loudness, instrumentalness, and tempo
}
\description{
Search for Spotify album track loudness and tempo
}
\examples{
\dontrun{
get_album_track_loudness_instrumentalness_tempo(id = "1uyf3l2d4XYwiEqAb7t7fX")
get_album_track_loudness_instrumentalness_tempo(query = "An Evening With Silk Sonic")
}
}
