% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album_functions.r
\name{get_albums}
\alias{get_albums}
\title{Search for Spotify album information}
\usage{
get_albums(
  queries = NULL,
  ids = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{queries}{\itemize{
\item A String vector of album names, taken from the search_spotify() function.  Defaults to NULL.  Only use this or ids
}}

\item{ids}{\itemize{
\item A vector of Spotify album ids.  Defaults to NULL.  Only use this or queries
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of album data, including album name and id, label, album popularity, release date, number of tracks, and artist info
}
\description{
Search for Spotify album information
}
\examples{
\dontrun{
get_albums(ids = "1uyf3l2d4XYwiEqAb7t7fX")
get_albums(ids = c("1uyf3l2d4XYwiEqAb7t7fX", "58ufpQsJ1DS5kq4hhzQDiI"))
get_albums(queries = c("An Evening With Silk Sonic", "To Pimp A Butterfly"))
}
}
