% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_functions.r
\name{get_albums_tracks}
\alias{get_albums_tracks}
\title{Search for Spotify album's tracks}
\usage{
get_albums_tracks(
  queries = NULL,
  ids = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{queries}{\itemize{
\item A vector of album names.  Deafualts to NULL.  Only use this or ids
}}

\item{ids}{\itemize{
\item A vector of Spotify album ids.  Deafualts to NULL.  Only use this or queries
}}

\item{limit}{\itemize{
\item Optional.  Number of albums wanted to return.  Valid if between 1 and 50.  Defaults to 20
}}

\item{offset}{\itemize{
\item Optional.  Index of first album wanted.  Defaults to 0
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of album track information, including duration, explicitness, track id and name, track number in album, and artist name and id
}
\description{
Search for Spotify album's tracks
}
\examples{
\dontrun{
get_albums_tracks(ids = c("4VZ7jhV0wHpoNPCB7Vmiml", "58ufpQsJ1DS5kq4hhzQDiI"))
get_albums_tracks(ids = "4VZ7jhV0wHpoNPCB7Vmiml", limit = 50, offset = 2)
get_albums_tracks(queries = "An Evening With Silk Sonic")
}
}
