% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album_functions.r
\name{get_artist_albums}
\alias{get_artist_albums}
\title{Search for Spotify artist's albums}
\usage{
get_artist_albums(
  query = NULL,
  id = NULL,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{query}{\itemize{
\item A single of artist name, taken from the search_spotify() function.  Defaults to NULL.  Only use this or id
}}

\item{id}{\itemize{
\item A single Spotify artist id.  Defaults to NULL.  Only use this or query
}}

\item{limit}{\itemize{
\item Optional.  Number of albums wanted to return.  Valid if between 1 and 50.  Defaults to 20
}}

\item{offset}{\itemize{
\item Optional.  Index of first album wanted.  Defaults to 0
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of album data, including ids and names for all albums by the artist, release date, number of tracks, and artists involved in the albums
}
\description{
Search for Spotify artist's albums
}
\examples{
\dontrun{
get_artist_albums(id = "0du5cEVh5yTK9QJze8zA0C")
get_artist_albums(id = "0du5cEVh5yTK9QJze8zA0C", limit = 50, offset = 2)
get_artist_albums(query = "Bruno Mars", limit = 44)
}
}
