% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist_functions.r
\name{get_artist_audio_features}
\alias{get_artist_audio_features}
\title{Search for Spotify artist audio feature information}
\usage{
get_artist_audio_features(
  query = NULL,
  id = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{query}{\itemize{
\item A single artist name.  Defaults to NULL.  Only put this or id
}}

\item{id}{\itemize{
\item A single Spotify artist id.  Defaults to NULL.  Only put this or query
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of artist track data, including the artist name and id, project id and name, release_date, track name and id, key and mode, and each of the following variables:
danceability
energy
loudness
speechiness
acousticness
instrumentalness
liveness
valence
tempo
duration_ms
}
\description{
Search for Spotify artist audio feature information
}
\examples{
\dontrun{
get_artist_audio_features(id = "0du5cEVh5yTK9QJze8zA0C")
get_artist_audio_features(query = "Bruno Mars")
}
}
