% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist_functions.r
\name{get_artists}
\alias{get_artists}
\title{Search for Spotify artist information}
\usage{
get_artists(
  queries = NULL,
  ids = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{queries}{\itemize{
\item A String vector of artist names, taken from the search_spotify() function.  Defaults to NULL.  Only use this or ids
}}

\item{ids}{\itemize{
\item A vector of Spotify artist ids.  Defaults to NULL.  Only use this or queries
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of artist data, including artist name and id, genres of the artist, artist popularity, and the total number of followers on spotify
}
\description{
Search for Spotify artist information
}
\examples{
\dontrun{
get_artists(ids = "0du5cEVh5yTK9QJze8zA0C")
get_artists(ids = c("0du5cEVh5yTK9QJze8zA0C", "6PvvGcCY2XtUcSRld1Wilr"))
get_artists(queries = c("Bruno Mars", "Anderson .Paak"))
}
}
