% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.r
\name{get_artists_danceability}
\alias{get_artists_danceability}
\title{Search for Spotify artists danceability summary}
\usage{
get_artists_danceability(
  queries = NULL,
  ids = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{queries}{\itemize{
\item A String vector of artist names, taken from the search_spotify() function.  Defaults to NULL.  Only use this or ids
}}

\item{ids}{\itemize{
\item A vector of Spotify artist ids.  Defaults to NULL.  Only use this or queries
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of artist summary data, including artist id, and the mean and standard deviation for danceability
}
\description{
Search for Spotify artists danceability summary
}
\examples{
\dontrun{
get_artists_danceability(ids = c("0du5cEVh5yTK9QJze8zA0C", "6PvvGcCY2XtUcSRld1Wilr"))
get_artists_danceability(queries = c("Bruno Mars", "Anderson .Paak"))
}
}
