% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.r
\name{get_artists_energy_num_songs}
\alias{get_artists_energy_num_songs}
\title{Search for Spotify artists energy summary and number of songs}
\usage{
get_artists_energy_num_songs(
  queries = NULL,
  ids = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{queries}{\itemize{
\item A String vector of artist names, taken from the search_spotify() function.  Defaults to NULL.  Only use this or ids
}}

\item{ids}{\itemize{
\item A vector of Spotify artist ids.  Defaults to NULL.  Only use this or queries
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of artist summary data, including artist name, number of songs, and mean energy
}
\description{
Search for Spotify artists energy summary and number of songs
}
\examples{
\dontrun{
get_artists_energy_num_songs(ids = c("0du5cEVh5yTK9QJze8zA0C", "6PvvGcCY2XtUcSRld1Wilr"))
get_artists_energy_num_songs(queries = c("Bruno Mars", "Anderson .Paak"))
}
}
