% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genre_functions.r
\name{get_genre_artists}
\alias{get_genre_artists}
\title{Search for Spotify artists of a specific genre}
\usage{
get_genre_artists(
  genre,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{genre}{\itemize{
\item A string of a genre
}}

\item{limit}{\itemize{
\item Optional.  Number of albums wanted to return.  Valid if between 1 and 50.  Defaults to 20
}}

\item{offset}{\itemize{
\item Optional.  Index of first album wanted.  Defaults to 0
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of artist data, including the artist's genre, artist name and id, artist popularity, and the total number of followers
}
\description{
Search for Spotify artists of a specific genre
}
\examples{
\dontrun{
get_genre_artists("hip-hop")
get_genre_artists("rap", limit = 50, offset = 2)
}
}
