% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genre_functions.r
\name{get_genre_summary}
\alias{get_genre_summary}
\title{Search for Spotify feature summary of tracks from a specific genre}
\usage{
get_genre_summary(genre, authorization = get_spotify_access_token())
}
\arguments{
\item{genre}{\itemize{
\item A string of a genre
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of genre track summary, including the inputted genre and the means and standard deviations for the following variables:
danceability
energy
loudness
speechiness
acousticness
instrumentalness
liveness
valence
tempo
duration_ms
mode
}
\description{
Search for Spotify feature summary of tracks from a specific genre
}
\examples{
\dontrun{
get_genre_summary("hip-hop")
}
}
