% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.r
\name{get_genre_tempo_speechiness_duration}
\alias{get_genre_tempo_speechiness_duration}
\title{Search for Spotify speechiness and duration of tracks from a specific genre}
\usage{
get_genre_tempo_speechiness_duration(
  genre = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{genre}{\itemize{
\item A string of a genre
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of genre track summary, including the inputted genre, mean tempo, mean speechiness, and mean duration in ms
}
\description{
Search for Spotify speechiness and duration of tracks from a specific genre
}
\examples{
\dontrun{
get_genre_key_speechiness_duration("hip-hop")
}
}
