% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genre_functions.r
\name{get_genre_track_features}
\alias{get_genre_track_features}
\title{Search for Spotify track features of a specific genre}
\usage{
get_genre_track_features(
  genre,
  limit = 20,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{genre}{\itemize{
\item A string of a genre
}}

\item{limit}{\itemize{
\item Optional.  Number of tracks wanted to return.  Valid if between 1 and 50.  Defaults to 20
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of track feature data, including explicitness, track name and id, track popularity, track number in its album, artist name and id, and each of the following variables:
danceability
energy
key
loudness
mode
speechiness
acousticness
instrumentalness
liveness
valence
tempo
duration_ms
time_signature
}
\description{
Search for Spotify track features of a specific genre
}
\examples{
\dontrun{
get_genre_track_features("hip-hop")
}
}
