% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genre_functions.r
\name{get_genre_tracks}
\alias{get_genre_tracks}
\title{Search for Spotify tracks of a specific genre}
\usage{
get_genre_tracks(
  genre,
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{genre}{\itemize{
\item A string of a genre
}}

\item{limit}{\itemize{
\item Optional.  Number of albums wanted to return.  Valid if between 1 and 50.  Defaults to 20
}}

\item{offset}{\itemize{
\item Optional.  Index of first album wanted.  Defaults to 0
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of track data, including duration, explicitness, track name and id, track popularity, whether the track was a single or on an album, and artist name and id
}
\description{
Search for Spotify tracks of a specific genre
}
\examples{
\dontrun{
get_genre_tracks("hip-hop")
get_genre_tracks("rap", limit = 50, offset = 2)
}
}
