% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.r
\name{get_new_releases}
\alias{get_new_releases}
\title{Search for Spotify new releases}
\usage{
get_new_releases(
  limit = 20,
  offset = 0,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{limit}{\itemize{
\item Optional.  Number of projects wanted to return.  Valid if between 1 and 50.  Defaults to 20
}}

\item{offset}{\itemize{
\item Optional.  Index of first project wanted.  Defaults to 0
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of new relase data, inlcuding the type, project name and id, release date, total number of tracks, and artist names and ids
}
\description{
Search for Spotify new releases
}
\examples{
\dontrun{
get_new_releases()
get_new_releases(limit = 50, offset = 2)
}
}
