% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genre_functions.r
\name{get_possible_genres}
\alias{get_possible_genres}
\title{Search for Spotify possible genres}
\usage{
get_possible_genres(authorization = get_spotify_access_token())
}
\arguments{
\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe containing all possible genre inputs for other genre functions
}
\description{
Search for Spotify possible genres
}
\examples{
\dontrun{
get_possible_genres()
}
}
