% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist_functions.r
\name{get_related_artists}
\alias{get_related_artists}
\title{Search for Spotify related artists}
\usage{
get_related_artists(
  query = NULL,
  id = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{query}{\itemize{
\item A single artist name.  Defaults to NULL.  Only use this or id
}}

\item{id}{\itemize{
\item A single Spotify artist id.  Defaults to NULL.  Only use this or id
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of related artist data, including the artist genres, artist id and name, popularity, and number of followers
}
\description{
Search for Spotify related artists
}
\examples{
\dontrun{
get_related_artists(id = "0du5cEVh5yTK9QJze8zA0C")
get_related_artists(query = "Bruno Mars")
}
}
