% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_functions.r
\name{get_spotify_access_token}
\alias{get_spotify_access_token}
\title{Get Spotify access token}
\usage{
get_spotify_access_token(
  client_id = Sys.getenv("SPOTIFY_CLIENT_ID"),
  client_secret = Sys.getenv("SPOTIFY_CLIENT_SECRET")
)
}
\arguments{
\item{client_id}{\itemize{
\item client id as given by the Spotify for Developers app.  Defaults to Sys.getenv('SPOTIFY_CLIENT_ID')
}}

\item{client_secret}{\itemize{
\item secret id as given by the Spotify for Developers app.  Defaults to Sys.getenv('SPOTIFY_CLIENT_SECRET')
}}
}
\value{
Spotify access token for the user
}
\description{
Get Spotify access token
}
\examples{
\dontrun{
get_spotify_access_token()
}
}
