% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_functions.r
\name{get_spotify_authorization_code}
\alias{get_spotify_authorization_code}
\title{Get Spotify authorization code}
\usage{
get_spotify_authorization_code(
  client_id = Sys.getenv("SPOTIFY_CLIENT_ID"),
  client_secret = Sys.getenv("SPOTIFY_CLIENT_SECRET"),
  scope = scopes()
)
}
\arguments{
\item{client_id}{\itemize{
\item client id as given by the Spotify for Developers app.  Defaults to Sys.getenv('SPOTIFY_CLIENT_ID')
}}

\item{client_secret}{\itemize{
\item secret id as given by the Spotify for Developers app.  Defaults to Sys.getenv('SPOTIFY_CLIENT_SECRET')
}}

\item{scope}{\itemize{
\item string of scopes for authorization access.  Defaults to all scopes
}}
}
\value{
Spotify authorization code for the user
}
\description{
Get Spotify authorization code
}
\examples{
\dontrun{
get_spotify_authorization_code()
}
}
