% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_functions.r
\name{get_track_analysis}
\alias{get_track_analysis}
\title{Search for Spotify track analysis}
\usage{
get_track_analysis(
  query = NULL,
  id = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{query}{\itemize{
\item A single track name.  Deafualts to NULL.  Only use this or id
}}

\item{id}{\itemize{
\item A single Spotify track id.  Deafualts to NULL.  Only use this or query
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of track data, including track id, duration, loudness, start and end of fade, tempo and confidence, time signature and confidence, key and confidence, and mode and confidence
}
\description{
Search for Spotify track analysis
}
\examples{
\dontrun{
get_track_analysis(id = "3w3y8KPTfNeOKPiqUTakBh")
get_track_analysis(query = "Locked out of Heaven")
}
}
