% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_functions.r
\name{get_track_audio_features}
\alias{get_track_audio_features}
\title{Search for Spotify track features}
\usage{
get_track_audio_features(
  queries = NULL,
  ids = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{queries}{\itemize{
\item A vector of track names.  Deafualts to NULL.  Only use this or ids
}}

\item{ids}{\itemize{
\item A vector of Spotify track ids.  Deafualts to NULL.  Only use this or queries
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of track feature information, including track id and the following variables:
danceability
energy
key
loudness
mode
speechiness
acousticness
instrumentalness
liveness
valence
tempo
duration_ms
time_signature
}
\description{
Search for Spotify track features
}
\examples{
\dontrun{
get_track_audio_features(ids = "3w3y8KPTfNeOKPiqUTakBh")
get_track_audio_features(queries = c("Locked out of Heaven", "Natalie"))
}
}
