% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_functions.r
\name{get_track_segments}
\alias{get_track_segments}
\title{Search for Spotify track segment information}
\usage{
get_track_segments(
  query = NULL,
  id = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{query}{\itemize{
\item A single track name.  Deafualts to NULL.  Only use this or id
}}

\item{id}{\itemize{
\item A single Spotify track id.  Deafualts to NULL.  Only use this or query
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of track segment information, including track id, start, end, and confidence of segment intervals, loudness at the start of the interval, time of maximum loudness, the maximum loudness, the end loudness, the pitches, and the timbre
}
\description{
Search for Spotify track segment information
}
\examples{
\dontrun{
get_track_segments(id = "3w3y8KPTfNeOKPiqUTakBh")
get_track_segments(query = "Locked out Heaven")
}
}
