% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_functions.r
\name{get_tracks}
\alias{get_tracks}
\title{Search for Spotify track information}
\usage{
get_tracks(
  queries = NULL,
  ids = NULL,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{queries}{\itemize{
\item A vector of track names.  Deafualts to NULL.  Only use this or ids
}}

\item{ids}{\itemize{
\item A vector of Spotify track ids.  Deafualts to NULL.  Only use this or queries
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of track information, including duration, explicitness, track name and id, popularity, whether the song was in an album or a single, album name and id, release date, number of tracks on album, and artist name and id
}
\description{
Search for Spotify track information
}
\examples{
\dontrun{
get_tracks(ids = "3G5iN5QBqMeXx3uZPy8tgB")
get_tracks(ids = c("3G5iN5QBqMeXx3uZPy8tgB", "3w3y8KPTfNeOKPiqUTakBh"))
get_tracks(queries = c("Locked out of Heaven", "When I Was Your Man"))
}
}
