% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\docType{data}
\name{pitch_class_lookup}
\alias{pitch_class_lookup}
\title{Search for Spotify pitches}
\format{
An object of class \code{character} of length 12.
}
\usage{
pitch_class_lookup
}
\value{
A vector of possible pitches
}
\description{
Search for Spotify pitches
}
\keyword{datasets}
