% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{scopes}
\alias{scopes}
\title{Search for Spotify possible scopes}
\usage{
scopes()
}
\value{
A vector of Spotify possible authorization scopes
}
\description{
Search for Spotify possible scopes
}
\examples{
\dontrun{
scopes()
}
}
