% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{search_spotify}
\alias{search_spotify}
\title{Gain information from the formal name of multiple artists, albums, or tracks}
\usage{
search_spotify(queries, type, authorization = get_spotify_access_token())
}
\arguments{
\item{queries}{\itemize{
\item A string vector of an artist names, album names, or track names
}}

\item{type}{\itemize{
\item A string of "artist", "album", or "track"
}}

\item{authorization}{\itemize{
\item An access_token generated from the get_spotify_access_token() function
}}
}
\value{
A dataframe of basic artist, track, or album information
}
\description{
Gain information from the formal name of multiple artists, albums, or tracks
}
\examples{
\dontrun{
 search_spotify(c("Bruno Mars", "Anderson .paak"), "artist")
 search_spotify(c("Natalie", "Grenade"), "track")
 search_spotify(c("An Evening With Silk Sonic", "Doo-Waps and Hooligans"), "album")
}
}
